FAD-5885 refactored render helpers for tests
