+++ src/__testHelpers__/index.js
@@ -1,3 +1,2 @@
 export { default as delay } from './delay';
 export { default as mockStore, createMockStore } from './mockStore';
-export { default as renderRowData } from './renderRowData';
