+++ src/__testHelpers__/renderHelpers.js
@@ -5,7 +5,10 @@ import { MemoryRouter, Link } from 'react-router-dom';
 
 const MockForm = reduxForm({ form: 'MockForm' })(({ children }) => <form>{children}</form>);
 
-function tryMount(item) {
+export const renderReduxFormField = (field) => shallow(<MockForm>{field}</MockForm>).find(Field);
+export const renderReactRouterLink = (link) => shallow(<MemoryRouter>{link}</MemoryRouter>).find(Link);
+
+export function tryMount(item) {
   let result = 'UNRENDERABLE';
   try {
     result = mount(item);
@@ -15,15 +18,15 @@ function tryMount(item) {
   return result;
 }
 
-function tryShallow(item) {
+export function tryShallow(item) {
   let result = 'UNRENDERABLE';
   try {
     result = shallow(item);
   } catch (err) {
     if (err.message === 'Field must be inside a component decorated with reduxForm()') {
-      result = shallow(<MockForm>{item}</MockForm>).find(Field);
+      result = renderReduxFormField(item);
     } else if (err.message === 'You should not use <Link> outside a <Router>') {
-      result = shallow(<MemoryRouter keyLength={0}>{item}</MemoryRouter>).find(Link);
+      result = renderReactRouterLink(item);
     } else {
       // eslint-disable-next-line no-console
       console.log('\n===\n\n UNKNOWN RENDER ROW DATA ERROR\n', err.message, '\n\n===');
@@ -39,6 +42,6 @@ function tryShallow(item) {
  *
  * @param {Array} row - list of items that may or may not be React elements
  */
-export default function renderRowData(row) {
+export function renderRowData(row) {
   return row.map((item) => React.isValidElement(item) ? tryMount(item) : item);
 }
