+++ src/pages/reports/bounce/tests/BouncePage.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BouncePage } from '../BouncePage';
 import { shallow, mount } from 'enzyme';
 import * as reportHelpers from 'src/helpers/reports';
-import { renderRowData } from 'src/__testHelpers__';
+import { renderRowData } from 'src/__testHelpers__/renderHelpers';
 import Filters from 'src/pages/reports/components/Filters';
 
 jest.mock('src/helpers/reports');
