+++ src/pages/reports/delays/tests/DelayPage.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { DelayPage } from '../DelayPage';
 import { shallow, mount } from 'enzyme';
 import * as reportHelpers from 'src/helpers/reports';
-import { renderRowData } from 'src/__testHelpers__';
+import { renderRowData } from 'src/__testHelpers__/renderHelpers';
 
 jest.mock('src/helpers/reports');
